requirements.txt
discord.py
yfinance
tensorflow-cpu
pandas
scikit-learn
joblib
numpy